LayoutTests/platform/chromium/media/video-capture-preview-expected.txt

request access to local media
got a stream
start preview
EVENT(loadstart)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
EVENT(play)
restart preview
EVENT(loadstart)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
EVENT(play)
END OF TEST
